Move functionality to appropriate file
