--- core/types/celo_extensions.go
@@ -1,15 +1,0 @@
-package types
-
-import (
-	"github.com/ethereum/go-ethereum/common"
-)
-
-// Returns the fee currency of the transaction if there is one.
-func (tx *Transaction) FeeCurrency() *common.Address {
-	var feeCurrency *common.Address
-	switch t := tx.inner.(type) {
-	case *CeloDynamicFeeTxV2:
-		feeCurrency = t.FeeCurrency
-	}
-	return feeCurrency
-}
